requirements.txt
future
emoji==0.6.0
requests
sqlalchemy==1.3.23
python-telegram-bot==11.1.0
psycopg2
feedparser
pythonping
pynewtonmath
#memes from SkittBot
spongemock
zalgo-text
geopy
nltk
aiohttp>=2.2.5
Pillow>=4.2.0
#Cool stuff by Peak
PyLyrics
#AEX
hurry.filesize
#translator
googletrans==3.1.0a0

tldextract
PyYAML
pyowm
wikipedia
bleach
markdown2
trakt.py
tmdbsimple
